+++ crates/opsml_app/src/app.rs
@@ -1,7 +1,7 @@
 use crate::error::AppError;
-use opsml_cards::CardDeck;
+use opsml_cards::ServiceCard;
 use opsml_state::app_state;
-use opsml_types::{cards::CardDeckMapping, SaveName, Suffix};
+use opsml_types::{cards::ServiceCardMapping, SaveName, Suffix};
 use pyo3::prelude::*;
 use pyo3::types::PyDict;
 use scouter_client::ScouterQueue;
@@ -9,17 +9,17 @@ use std::path::{Path, PathBuf};
 use tracing::{debug, error};
 
 /// Load a card map from path
-fn load_card_map(path: &Path) -> Result<CardDeckMapping, AppError> {
+fn load_card_map(path: &Path) -> Result<ServiceCardMapping, AppError> {
     let card_mapping_path = path.join(SaveName::CardMap).with_extension(Suffix::Json);
     debug!("Loading card mapping from: {:?}", card_mapping_path);
-    let mapping = CardDeckMapping::from_path(&card_mapping_path)?;
+    let mapping = ServiceCardMapping::from_path(&card_mapping_path)?;
     Ok(mapping)
 }
 
 #[pyclass]
 #[derive(Debug)]
 pub struct AppState {
-    deck: Py<CardDeck>,
+    service: Py<ServiceCard>,
     queue: Option<Py<ScouterQueue>>,
 }
 
@@ -28,31 +28,31 @@ impl AppState {
     /// Instantiate a new application state. Typically from_path is used; however, an new/init
     /// method is provided for consistency with other classes
     /// # Arguments
-    /// * `deck` - The CardDeck to use for the application state
+    /// * `service` - The ServiceCard to use for the application state
     /// * `queue` - An optional ScouterQueue to use for the application state. If not provided,
     /// no queue will be created.
     ///
     /// # Returns
-    /// * `AppState` - The application state containing the CardDeck and optional ScouterQueue  
+    /// * `AppState` - The application state containing the ServiceCard and optional ScouterQueue  
     #[new]
-    #[pyo3(signature = (deck, queue=None))]
+    #[pyo3(signature = (service, queue=None))]
     pub fn new(
-        deck: Bound<'_, CardDeck>,
+        service: Bound<'_, ServiceCard>,
         queue: Option<Bound<'_, ScouterQueue>>,
     ) -> Result<Self, AppError> {
-        let deck = deck.unbind();
+        let service = service.unbind();
         let queue = queue.map(|q| q.unbind());
-        Ok(AppState { deck, queue })
+        Ok(AppState { service, queue })
     }
     /// This method will load an application state from a path
     /// This is primarily used for loading an application during api start where a user
-    /// may wish to load an Opsml CardDeck along with the appropriate ScouterQueue for monitoring
+    /// may wish to load an Opsml ServiceCard along with the appropriate ScouterQueue for monitoring
     /// This class and it's functionality may expand in the future
     ///
     /// # Arguments
     /// * `py` - Python interpreter state
     /// * `path` - The root path to the application directory containing files
-    /// * `load_kwargs` - Load kwargs to pass to the CardDeck loader
+    /// * `load_kwargs` - Load kwargs to pass to the ServiceCard loader
     /// * `transport_config` = The transport config to use with the ScouterQueue. If not provided,
     /// no queue will be created.
     #[staticmethod]
@@ -63,8 +63,8 @@ impl AppState {
         load_kwargs: Option<&Bound<'_, PyDict>>,
         transport_config: Option<&Bound<'_, PyAny>>,
     ) -> Result<Self, AppError> {
-        let path = path.unwrap_or_else(|| PathBuf::from(SaveName::CardDeck));
-        let deck = Py::new(py, CardDeck::from_path_rs(py, &path, load_kwargs)?)?;
+        let path = path.unwrap_or_else(|| PathBuf::from(SaveName::ServiceCard));
+        let service = Py::new(py, ServiceCard::from_path_rs(py, &path, load_kwargs)?)?;
         let card_map = load_card_map(&path).map_err(|e| {
             error!("Failed to load card map from: {:?}", e);
             e
@@ -89,12 +89,12 @@ impl AppState {
             None
         };
 
-        Ok(AppState { deck, queue })
+        Ok(AppState { service, queue })
     }
 
     #[getter]
-    pub fn deck<'py>(&self, py: Python<'py>) -> Result<&Bound<'py, CardDeck>, AppError> {
-        Ok(self.deck.bind(py))
+    pub fn service<'py>(&self, py: Python<'py>) -> Result<&Bound<'py, ServiceCard>, AppError> {
+        Ok(self.service.bind(py))
     }
 
     #[getter]