+++ crates/opsml_cards/src/lib.rs
@@ -1,15 +1,15 @@
 pub mod data;
-pub mod deck;
 pub mod error;
 pub mod experiment;
 pub mod model;
 pub mod prompt;
+pub mod service;
 pub mod traits;
 pub mod utils;
 
 pub use data::*;
-pub use deck::*;
 pub use experiment::*;
 pub use model::*;
 pub use prompt::*;
+pub use service::*;
 pub use utils::*;