+++ crates/opsml_registry/src/registry.rs
@@ -422,7 +422,7 @@ impl CardRegistry {
         let tmp_path = tmp_dir.keep();
 
         match registry_type {
-            RegistryType::Experiment | RegistryType::Prompt | RegistryType::Deck => {
+            RegistryType::Experiment | RegistryType::Prompt | RegistryType::Service => {
                 card.call_method1("save", (tmp_path.to_path_buf(),))
                     .inspect_err(|e| {
                         error!("Failed to save card: {}", e);
@@ -543,7 +543,7 @@ impl CardRegistry {
         let tmp_path = tmp_dir.keep();
 
         match registry_type {
-            RegistryType::Experiment | RegistryType::Deck => {
+            RegistryType::Experiment | RegistryType::Service => {
                 card.call_method1("save", (tmp_path.to_path_buf(),))
                     .inspect_err(|e| {
                         error!("Failed to save card: {}", e);
@@ -803,7 +803,7 @@ pub struct CardRegistries {
     pub prompt: CardRegistry,
 
     #[pyo3(get)]
-    pub deck: CardRegistry,
+    pub service: CardRegistry,
 }
 
 #[pymethods]
@@ -821,14 +821,14 @@ impl CardRegistries {
         let model = CardRegistry::rust_new(&RegistryType::Model)?;
         let data = CardRegistry::rust_new(&RegistryType::Data)?;
         let prompt = CardRegistry::rust_new(&RegistryType::Prompt)?;
-        let deck = CardRegistry::rust_new(&RegistryType::Deck)?;
+        let service = CardRegistry::rust_new(&RegistryType::Service)?;
 
         Ok(Self {
             experiment,
             model,
             data,
             prompt,
-            deck,
+            service,
         })
     }
 }