+++ crates/opsml_server/opsml_ui/src/routes/opsml/service/card/home/+page.svelte
@@ -1,17 +1,14 @@
 <script lang="ts">
-  // version $props() in +page.svelte
-
-  import { goto } from '$app/navigation';
-  import type { CardDeck } from '$lib/components/card/card_interfaces/carddeck';
+  import type { ServiceCard } from '$lib/components/card/card_interfaces/servicecard';
   import type { PageProps } from './$types';
-  import DeckMetadata from '$lib/components/card/deck/DeckMetadata.svelte';
+  import ServiceMetadata from '$lib/components/card/service/ServiceMetadata.svelte';
   import { Table } from 'lucide-svelte';
-  import CardButton from '$lib/components/card/deck/CardButton.svelte';
+  import CardButton from '$lib/components/card/service/CardButton.svelte';
   let badgeColor = "#40328b";
   let iconColor = "#40328b";
 
   let { data }: PageProps = $props();
-  let deck: CardDeck = data.metadata;
+  let service: ServiceCard = data.metadata;
 
 </script>
 
@@ -21,7 +18,7 @@
   <div class="flex flex-wrap pt-4 gap-4 w-full justify-center">
 
     <div class="bg-primary-200 p-4 flex-1 flex-col rounded-base bg-surface-50 border-primary-800 border-3 shadow-primary max-h-[800px] overflow-y-auto self-start min-w-[26rem] max-w-[300px] md:min-w-[26rem] md:max-w-[32rem]">
-        <DeckMetadata deck={deck}/>
+        <ServiceMetadata service={service}/>
     </div>
 
     <div class="flex-1 flex-col rounded-base border-primary-500 border-2 shadow-primary bg-surface-50 pb-4 px-4">
@@ -35,7 +32,7 @@
       <div class="grid grid-cols-1 lg:grid-cols-2 gap-4">
      
 
-          {#each deck.cards.cards as card}
+          {#each service.cards.cards as card}
             <CardButton
               iconColor={iconColor} 
               badgeColor={badgeColor}