+++ crates/opsml_sql/src/mysql/sql/card/insert_servicecard.sql
@@ -1,0 +1,1 @@
+INSERT INTO opsml_service_registry (uid, app_env, name, space, major, minor, patch, version, pre_tag, build_tag, cards, username, opsml_version) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?);