+++ crates/opsml_sql/src/postgres/client.rs
@@ -3,9 +3,9 @@ use crate::base::SqlClient;
 use crate::error::SqlError;
 use crate::postgres::helper::PostgresQueryHelper;
 use crate::schemas::schema::{
-    AuditCardRecord, CardDeckRecord, CardResults, CardSummary, DataCardRecord,
-    ExperimentCardRecord, HardwareMetricsRecord, MetricRecord, ModelCardRecord, ParameterRecord,
-    PromptCardRecord, QueryStats, ServerCard, SqlSpaceRecord, User, VersionResult, VersionSummary,
+    AuditCardRecord, CardResults, CardSummary, DataCardRecord, ExperimentCardRecord,
+    HardwareMetricsRecord, MetricRecord, ModelCardRecord, ParameterRecord, PromptCardRecord,
+    QueryStats, ServerCard, ServiceCardRecord, SqlSpaceRecord, User, VersionResult, VersionSummary,
 };
 use async_trait::async_trait;
 use opsml_semver::VersionValidator;
@@ -236,8 +236,8 @@ impl SqlClient for PostgresClient {
                 return Ok(CardResults::Prompt(card));
             }
 
-            CardTable::Deck => {
-                let card: Vec<CardDeckRecord> = sqlx::query_as(&query)
+            CardTable::Service => {
+                let card: Vec<ServiceCardRecord> = sqlx::query_as(&query)
                     .bind(query_args.uid.as_ref())
                     .bind(query_args.space.as_ref())
                     .bind(query_args.name.as_ref())
@@ -246,7 +246,7 @@ impl SqlClient for PostgresClient {
                     .fetch_all(&self.pool)
                     .await?;
 
-                return Ok(CardResults::Deck(card));
+                return Ok(CardResults::Service(card));
             }
             _ => {
                 return Err(SqlError::InvalidTableName);
@@ -332,7 +332,7 @@ impl SqlClient for PostgresClient {
                         .bind(&record.datacard_uids)
                         .bind(&record.modelcard_uids)
                         .bind(&record.promptcard_uids)
-                        .bind(&record.card_deck_uids)
+                        .bind(&record.service_card_uids)
                         .bind(&record.experimentcard_uids)
                         .bind(&record.pre_tag)
                         .bind(&record.build_tag)
@@ -405,9 +405,9 @@ impl SqlClient for PostgresClient {
                     return Err(SqlError::InvalidCardType);
                 }
             },
-            CardTable::Deck => match card {
-                ServerCard::Deck(record) => {
-                    let query = PostgresQueryHelper::get_carddeck_insert_query();
+            CardTable::Service => match card {
+                ServerCard::Service(record) => {
+                    let query = PostgresQueryHelper::get_servicecard_insert_query();
                     sqlx::query(&query)
                         .bind(&record.uid)
                         .bind(&record.app_env)
@@ -515,7 +515,7 @@ impl SqlClient for PostgresClient {
                         .bind(&record.datacard_uids)
                         .bind(&record.modelcard_uids)
                         .bind(&record.promptcard_uids)
-                        .bind(&record.card_deck_uids)
+                        .bind(&record.service_card_uids)
                         .bind(&record.experimentcard_uids)
                         .bind(&record.pre_tag)
                         .bind(&record.build_tag)
@@ -589,9 +589,9 @@ impl SqlClient for PostgresClient {
                 }
             },
 
-            CardTable::Deck => match card {
-                ServerCard::Deck(record) => {
-                    let query = PostgresQueryHelper::get_carddeck_update_query();
+            CardTable::Service => match card {
+                ServerCard::Service(record) => {
+                    let query = PostgresQueryHelper::get_servicecard_update_query();
                     sqlx::query(&query)
                         .bind(&record.app_env)
                         .bind(&record.name)
@@ -1201,7 +1201,7 @@ impl SqlClient for PostgresClient {
 
 #[cfg(test)]
 mod tests {
-    use crate::schemas::CardDeckRecord;
+    use crate::schemas::ServiceCardRecord;
 
     use super::*;
     use opsml_types::{CommonKwargs, RegistryType, SqlType};
@@ -1244,7 +1244,7 @@ mod tests {
             FROM opsml_audit_event;
 
             DELETE
-            FROM opsml_deck_registry;
+            FROM opsml_service_registry;
 
             DELETE
             FROM opsml_space;
@@ -1267,7 +1267,7 @@ mod tests {
             CardTable::Experiment => ServerCard::Experiment(ExperimentCardRecord::default()),
             CardTable::Audit => ServerCard::Audit(AuditCardRecord::default()),
             CardTable::Prompt => ServerCard::Prompt(PromptCardRecord::default()),
-            CardTable::Deck => ServerCard::Deck(CardDeckRecord::default()),
+            CardTable::Service => ServerCard::Service(ServiceCardRecord::default()),
             _ => panic!("Invalid card type"),
         };
 
@@ -1278,7 +1278,7 @@ mod tests {
             ServerCard::Experiment(c) => c.uid.clone(),
             ServerCard::Audit(c) => c.uid.clone(),
             ServerCard::Prompt(c) => c.uid.clone(),
-            ServerCard::Deck(c) => c.uid.clone(),
+            ServerCard::Service(c) => c.uid.clone(),
         };
 
         // Test Insert
@@ -1335,13 +1335,13 @@ mod tests {
                 };
                 ServerCard::Prompt(c)
             }
-            CardTable::Deck => {
-                let c = CardDeckRecord {
+            CardTable::Service => {
+                let c = ServiceCardRecord {
                     uid: uid.clone(),
                     name: updated_name.to_string(),
                     ..Default::default()
                 };
-                ServerCard::Deck(c)
+                ServerCard::Service(c)
             }
             _ => panic!("Invalid card type"),
         };
@@ -1360,7 +1360,7 @@ mod tests {
             CardResults::Experiment(cards) => assert_eq!(cards[0].name, updated_name),
             CardResults::Audit(cards) => assert_eq!(cards[0].name, updated_name),
             CardResults::Prompt(cards) => assert_eq!(cards[0].name, updated_name),
-            CardResults::Deck(cards) => assert_eq!(cards[0].name, updated_name),
+            CardResults::Service(cards) => assert_eq!(cards[0].name, updated_name),
         }
 
         // delete card
@@ -1599,7 +1599,7 @@ mod tests {
         test_card_crud(&client, &CardTable::Prompt, "UpdatedPromptName")
             .await
             .unwrap();
-        test_card_crud(&client, &CardTable::Deck, "UpdatedDeckName")
+        test_card_crud(&client, &CardTable::Service, "UpdatedDeckName")
             .await
             .unwrap();
     }