+++ crates/opsml_sql/src/sqlite/helper.rs
@@ -38,13 +38,13 @@ const INSERT_PROMPTCARD_SQL: &str = include_str!("sql/card/insert_promptcard.sql
 const INSERT_MODELCARD_SQL: &str = include_str!("sql/card/insert_modelcard.sql");
 const INSERT_EXPERIMENTCARD_SQL: &str = include_str!("sql/card/insert_experimentcard.sql");
 const INSERT_AUDITCARD_SQL: &str = include_str!("sql/card/insert_auditcard.sql");
-const INSERT_CARDDECK_SQL: &str = include_str!("sql/card/insert_carddeck.sql");
+const INSERT_SERVICECARD_SQL: &str = include_str!("sql/card/insert_servicecard.sql");
 const UPDATE_DATACARD_SQL: &str = include_str!("sql/card/update_datacard.sql");
 const UPDATE_PROMPTCARD_SQL: &str = include_str!("sql/card/update_promptcard.sql");
 const UPDATE_MODELCARD_SQL: &str = include_str!("sql/card/update_modelcard.sql");
 const UPDATE_EXPERIMENTCARD_SQL: &str = include_str!("sql/card/update_experimentcard.sql");
 const UPDATE_AUDITCARD_SQL: &str = include_str!("sql/card/update_auditcard.sql");
-const UPDATE_CARDDECK_SQL: &str = include_str!("sql/card/update_carddeck.sql");
+const UPDATE_SERVICECARD_SQL: &str = include_str!("sql/card/update_servicecard.sql");
 
 // artifact keys
 const INSERT_ARTIFACT_KEY_SQL: &str = include_str!("sql/artifact/insert_artifact_key.sql");
@@ -433,12 +433,12 @@ impl SqliteQueryHelper {
         INSERT_AUDITCARD_SQL.to_string()
     }
 
-    pub fn get_carddeck_insert_query() -> String {
-        INSERT_CARDDECK_SQL.to_string()
+    pub fn get_servicecard_insert_query() -> String {
+        INSERT_SERVICECARD_SQL.to_string()
     }
 
-    pub fn get_carddeck_update_query() -> String {
-        UPDATE_CARDDECK_SQL.to_string()
+    pub fn get_servicecard_update_query() -> String {
+        UPDATE_SERVICECARD_SQL.to_string()
     }
 
     pub fn get_promptcard_update_query() -> String {